Extract worker output prefix state
